metadata: optimize FieldAccessor callbacks to check for __napi::this property
Small optimization allows to quickly check if `this` in field setter/getter calls has a valid java object linked to it. The slow path is really slow, as it requires entering back into JS multiple times. On the other hand, just checking the property exists is really fast. Still not as fast as v8's built in `info->Holder() == info->This()` but we get close enough.